Fix redirect back to dashboard when products are purged

+++ app/Http/Controllers/SettingsController.php
@@ -28,9 +28,11 @@ class SettingsController extends Controller
         // If request to purge all products, do that noew
         if ($request->input('type') === 'purge_all_products') {
             Product::truncate();
-            chdir(storage_path(config('nwwsoi-controller.nwwsoi.archivedir')));
-            exec('rm -rf *');
-            return Redirect::route('settings.edit')->with('status', 'All products have been deleted!');
+            if (!is_null(config('nwwsoi-controller.nwwsoi.archivedir')) && !empty(config('nwwsoi-controller.nwwsoi.archivedir'))) {
+                chdir(storage_path(config('emwin-controller.emwin.archivedir')));
+                exec('rm -rf *');
+            }
+            return Redirect::route('dashboard');
         }
         return Redirect::route('settings.edit')->with('status', 'settings-updated');
     }
